Audio/ValKit: Factored out cleaning up the registered tests in the ValKit audio driver and also do so when ending a test set [build fix]. bugref:10008
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@89838 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Devices/Audio/DrvHostAudioValidationKit.cpp
@@ -228,7 +228,7 @@ static void drvHostValKitCleanup(PDRVHOSTVALKITAUDIO pThis)
         size_t const cbOutstanding = pTst->t.TestTone.u.Rec.cbToWrite - pTst->t.TestTone.u.Rec.cbWritten;
         if (cbOutstanding)
             LogRel(("Audio: Validation Kit: \tRecording test #%RU32 has %RU64 bytes (%RU32ms) outstanding\n",
-                    pTst->idxTest, cbOutstanding, PDMAudioPropsBytesToMilli(&pTst->t.TestTone.Parms.Props, cbOutstanding)));
+                    pTst->idxTest, cbOutstanding, PDMAudioPropsBytesToMilli(&pTst->t.TestTone.Parms.Props, (uint32_t)cbOutstanding)));
         drvHostValKiUnregisterRecTest(pThis, pTst);
     }
 
@@ -240,7 +240,7 @@ static void drvHostValKitCleanup(PDRVHOSTVALKITAUDIO pThis)
         size_t const cbOutstanding = pTst->t.TestTone.u.Play.cbToRead - pTst->t.TestTone.u.Play.cbRead;
         if (cbOutstanding)
             LogRel(("Audio: Validation Kit: \tPlayback test #%RU32 has %RU64 bytes (%RU32ms) outstanding\n",
-                    pTst->idxTest, cbOutstanding, PDMAudioPropsBytesToMilli(&pTst->t.TestTone.Parms.Props, cbOutstanding)));
+                    pTst->idxTest, cbOutstanding, PDMAudioPropsBytesToMilli(&pTst->t.TestTone.Parms.Props, (uint32_t)cbOutstanding)));
         drvHostValKiUnregisterPlayTest(pThis, pTst);
     }
